Updated GameArena and Arrow to support addArrow
New functions in gameArena:
+ public void addArrow
+ public void removeArrow
Both function in the same way as addLine and removeLine but take an Arrow as their argument.

Removed old non-object oriented methods i.e. including GameArena in constructor, from class Arrow
- Changed Arrow constructor

Added new public function to Arrow to get the three lines that make it up.

This method is much more clear and follows previous conventions of the GameArena class and the classes that can be used in it (such as Rectangle, Line etc.) and allows much easier management of arrows in the GameArena environment.